Renaming component to avoid collision
