+++ lib/index.js
@@ -11,7 +11,7 @@ import OnlyClickListOptions from './components/OnlyClickListOptions';
 import OnlyClickIconOption from './components/OnlyClickIconOption';
 import OnlyClickIconOptions from './components/OnlyClickIconOptions';
 import OnlyClickSelect from './components/OnlyClickSelect';
-import FlashNotification from './components/FlashNotification';
+import CwFlashNotification from './components/CwFlashNotification';
 
 export {
   Loader,
@@ -27,5 +27,5 @@ export {
   OnlyClickListOptions,
   OnlyClickIconOption,
   OnlyClickIconOptions,
-  FlashNotification,
+  CwFlashNotification,
 };
