+++ lib/stylesheets/components/_components.scss
@@ -10,4 +10,4 @@
 @import 'oc_icon_options';
 @import 'oc_selected_value';
 @import 'oc_select';
-@import 'flash_notification';
+@import 'cw_flash_notification';
